pyproject.toml
[project]
name = "multiai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "crewai>=1.6.1",
    "faiss-cpu>=1.13.2",
    "openai>=2.6.1",
    "sentence-transformers>=5.2.0",
]
